feat: 🎸 new modal for global setting
