+++ .gitignore
@@ -28,4 +28,5 @@ dist-ssr
 #local
 .history
 .local
+yarn.lock
