+++ src/assets/css/index.css
@@ -5,3 +5,40 @@
 [v-cloak] {
 	display: none;
 }
+
+/* 滚动条样式 */
+::-webkit-scrollbar-track {
+	border-radius: 10px;
+}
+
+::-webkit-scrollbar-thumb {
+	background-color: rgba(144, 147, 153, 0.3);
+	border-radius: 10px;
+}
+
+::-webkit-scrollbar {
+	width: 8px;
+	height: 16px;
+}
+
+::-webkit-scrollbar-track,
+::-webkit-scrollbar-thumb {
+	border: 5px solid transparent;
+	border-radius: 4px;
+}
+
+::-webkit-scrollbar-thumb {
+	min-height: 16px;
+	background-clip: content-box;
+	border: 1px solid rgba(128, 128, 128, 0.8);
+}
+
+::-webkit-scrollbar-corner {
+	background: transparent;
+}
+
+html::-webkit-scrollbar,
+body::-webkit-scrollbar,
+#root::-webkit-scrollbar {
+	width: 0;
+}
