+++ src/assets/css/setting-box.less
@@ -1,0 +1,69 @@
+.setting-box {
+	.arco-menu-vertical {
+		font-size: 12px;
+	}
+
+	.arco-menu-vertical .arco-menu-inner {
+		padding: 16px 8px;
+	}
+
+	.arco-menu-item {
+		line-height: 32px;
+		color: #333333;
+	}
+	.arco-menu-inline-header {
+		color: #333333;
+		line-height: 32px;
+		font-weight: 600;
+		pointer-events: none;
+		cursor: default;
+	}
+	.arco-menu-inline {
+		margin-bottom: 12px;
+	}
+
+	.arco-menu-light .arco-menu-inline-header.arco-menu-selected {
+		color: #333333;
+		font-weight: 600;
+	}
+
+	.arco-modal-body {
+		padding: 0;
+	}
+	.arco-menu-icon-suffix {
+		display: none;
+	}
+	.arco-layout-sider {
+		border-radius: 16px;
+	}
+
+	.label-box {
+		padding: 8px 0;
+		color: #333333;
+		.arco-form-item {
+			margin: 0;
+		}
+		border-top: 1px solid #c6c6c6;
+		.arco-form-item-content-flex {
+			justify-content: flex-end;
+		}
+	}
+
+	.layout-content {
+		display: flex;
+		flex-direction: column;
+		.close-icon {
+			text-align: right;
+			.arco-btn-shape-circle {
+				color: #333333;
+				height: 24px;
+				width: 24px;
+			}
+		}
+		.layout-box {
+			padding: 0 32px 16px 32px;
+			flex-grow: 1;
+			overflow-y: auto;
+		}
+	}
+}
